+++ pkg/bwagreement/bwagreement_test/common.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package testbwagreement
+package bwagreement_test
 
 import (
 	"time"
